Move router config directories into configs/.
