+++ configs/finland/gtfs-rules/hsl-no-trains.rule
