+++ configs/hsl/gtfs-rules/remove-route-color.rule
