+++ configs/hsl/osm-preprocessing/hsl.sh
