+++ configs/kela/gtfs-rules/remove-route-color.rule
