+++ configs/varely/config.js
@@ -1,4 +1,4 @@
-const mapSrc = require('../utils/configUtils.js');
+const mapSrc = require('../../utils/configUtils.js');
 
 module.exports = {
   id: 'varely',
