+++ configs/waltti/gtfs-rules/only-kotka-ferries.rule
